Heroku

+++ commands/resume.js
@@ -5,7 +5,9 @@ module.exports = {
         if(dispatcher)
         {
             dispatcher.resume();
-            message.delete();
+
+            if(message.content)
+                message.delete();
         }
 	},
 };
